refactor(event-item): extract drag and resize logic into custom hooks
Extracted drag and resize logic from EventItem component into separate custom hooks (useEventDrag and useEventResize) to improve code maintainability and reusability. Also moved time utility functions to a dedicated utils file. This refactor reduces complexity in the EventItem component and makes the drag/resize behavior more modular and testable.
